+++ src/main/java/com/example/demo/services/ManualAssetPriceService.java
@@ -5,6 +5,7 @@ import com.example.demo.infrastructure.repository.Asset;
 import com.example.demo.infrastructure.repository.AssetRepository;
 import com.example.demo.infrastructure.repository.AssetPriceHistoryRepository;
 import com.example.demo.services.currency.CurrencyRateService;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.ZonedDateTime;
 import lombok.RequiredArgsConstructor;
@@ -59,7 +60,7 @@ public class ManualAssetPriceService {
         asset.getSymbol(),
         "MANUAL",
         currency.name(),
-        marketPrice,
+        BigDecimal.valueOf(marketPrice),
         100,
         "MANUAL");
